// frontend/tina/__generated__/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** References another document, used as a foreign key */
  Reference: { input: any; output: any; }
  JSON: { input: any; output: any; }
};
// ...
export type PageBlocksVersionsMainCards = {
  __typename?: 'PageBlocksVersionsMainCards';
  title?: Maybe<Scalars['String']['output']>;
  image?: Maybe<Scalars['String']['output']>;
};

export type PageBlocksVersionsVersionFeatures = {
  __typename?: 'PageBlocksVersionsVersionFeatures';
  title?: Maybe<Scalars['String']['output']>;
  content?: Maybe<Scalars['String']['output']>;
  icon?: Maybe<Scalars['String']['output']>;
};

export type PageBlocksVersions = {
  __typename?: 'PageBlocksVersions';
  titleLine1?: Maybe<Scalars['String']['output']>;
  titleLine2?: Maybe<Scalars['String']['output']>;
  mainCards?: Maybe<Array<Maybe<PageBlocksVersionsMainCards>>>;
  versionFeatures?: Maybe<Array<Maybe<PageBlocksVersionsVersionFeatures>>>;
};